fixed negative hex conversion by casting argument to unsigned int

+++ arg_conversions/print_hex_low.c
@@ -33,7 +33,7 @@ int	print_hex_low(int n)
 	int		count;
 	char	*result;
 
-	result = to_lower(print_hexadecimal(n));
+	result = to_lower(print_hexadecimal((unsigned int)n));
 	count = write_str(result);
 	return (count);
 }
